+++ src/plugins/liveSearch/edit/editLiveSearch.interface.ts
@@ -35,6 +35,11 @@ export interface EditLiveSearchOptions extends LiveSearchOptions<CssClassesEditL
      * Indication whether use non existing value in live search as new value, works for multi value and only with dynamic value handler
      */
     useNonExistingAsValue?: boolean;
+
+    /**
+     * Min length of input that will filter down available options
+     */
+    minLengthSearch?: number;
 }
 
 /**